# lib/api-zod/src/generated/types/createPropertyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Board Game Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePropertyBody {
  name: string;
  dataType: string;
  /** @nullable */
  defaultValue?: string | null;
  /** @nullable */
  minValue?: string | null;
  /** @nullable */
  maxValue?: string | null;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  description?: string | null;
}