# lib/api-zod/src/generated/types/projectStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Board Game Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectStats {
  projectId: number;
  entityCount: number;
  propertyCount: number;
  ruleCount: number;
  assetCount: number;
  simulationCount: number;
}