# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Board Game Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  genre?: string | null;
  /** @nullable */
  playerCount?: string | null;
  /** @nullable */
  targetDuration?: string | null;
}